✨ feat(screen): remodel app layout from stack to tab navigation
- create new HomeLayout in tk-capture-book app
- refactor usage of react hooks in BookCaptureScreen
- move book-capture, book-chat, book-detail, book-search-detail to new home directory
- remove old tab layout and relocate to main auth layout
- update router paths to reflect structure change
- adjust styling parameters in book-detail-capture and book-detail-chat components
- change route paths in other files to match new structure
