+++ apps/tk-capture-book/app/(auth)/(home)/book-detail.tsx
