+++ apps/tk-capture-book/app/(auth)/(home)/book-search-detail.tsx
