--- apps/tk-capture-book/app/(auth)/(tabs)/_layout.tsx
@@ -1,73 +1,0 @@
-import { Ionicons } from "@expo/vector-icons";
-import { Tabs } from "expo-router";
-import { useTranslation } from "react-i18next";
-
-export default function TabLayout() {
-  const { t } = useTranslation();
-
-  return (
-    <Tabs
-      screenOptions={{
-        tabBarActiveTintColor: "#0284c7",
-        tabBarInactiveTintColor: "#64748b",
-        tabBarIconStyle: {
-          marginBottom: -4,
-        },
-        tabBarLabelStyle: {
-          marginTop: 4,
-          fontSize: 12,
-        },
-      }}
-    >
-      <Tabs.Screen
-        name="index"
-        options={{
-          title: t("home.tabTitle"),
-          tabBarIcon: ({ color, size, focused }) => (
-            <Ionicons name={focused ? "home" : "home-outline"} size={size} color={color} />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="frame"
-        options={{
-          title: t("frame.tabTitle"),
-          tabBarIcon: ({ color, size, focused }) => (
-            <Ionicons name={focused ? "film" : "film-outline"} size={size} color={color} />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="camera"
-        options={{
-          title: t("collectSentence.tabTitle"),
-          tabBarIcon: ({ color, size, focused }) => (
-            <Ionicons name={focused ? "list" : "list-outline"} size={size} color={color} />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="book-talk"
-        options={{
-          title: t("bookTalk.tabTitle"),
-          tabBarIcon: ({ color, size, focused }) => (
-            <Ionicons
-              name={focused ? "chatbubble-ellipses" : "chatbubble-ellipses-outline"}
-              size={size}
-              color={color}
-            />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="settings"
-        options={{
-          title: t("settings.tabTitle"),
-          tabBarIcon: ({ color, size, focused }) => (
-            <Ionicons name={focused ? "settings" : "settings-outline"} size={size} color={color} />
-          ),
-        }}
-      />
-    </Tabs>
-  );
-}
